Integration tests were added
