+++ .vscode/launch.json
@@ -1,0 +1,21 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "Run Debug",
+            "type": "go",
+            "request": "launch",
+            "mode": "debug",
+            "cwd": "${workspaceRoot}",
+            "program": "cmd/dcgm-exporter/main.go",
+            "args": [
+                "-f",
+                "./etc/default-counters.csv",
+                "--web-config-file=./tests/integration/testdata/web-config.yml"
+            ]
+        }
+    ]
+}
