+++ staticcheck.conf
@@ -1,0 +1,1 @@
+checks = ["all", "-ST1005"]
