+++ tests/integration/start_with_tls_test.go
@@ -1,0 +1,96 @@
+package integration
+
+import (
+	"context"
+	"crypto/tls"
+	"fmt"
+	"net/http"
+	"os"
+	"testing"
+	"time"
+
+	"github.com/NVIDIA/dcgm-exporter/pkg/cmd"
+	"github.com/avast/retry-go/v4"
+	"github.com/stretchr/testify/require"
+)
+
+func TestStartWithTLSEnabledAndBasicAuth(t *testing.T) {
+	app := cmd.NewApp()
+	args := os.Args[0:1]
+	args = append(args, "-f=../../etc/default-counters.csv") // Append a file with default counters
+	port := getRandomAvailablePort(t)
+	args = append(args, fmt.Sprintf("-a=:%d", port))
+	args = append(args, "--web-config-file=./testdata/web-config.yml")
+	ctx, cancel := context.WithCancel(context.Background())
+	go func(ctx context.Context) {
+		app.Run(args)
+	}(ctx)
+
+	t.Run("server returns 400 if request uses HTTP and TLS enabled on the server",
+		func(t *testing.T) {
+			status, err := retry.DoWithData(
+				func() (int, error) {
+					_, status, err := httpGet(t, fmt.Sprintf("http://localhost:%d/metrics", port))
+					if err != nil {
+						return -1, err
+					}
+					return status, nil
+				},
+				retry.Attempts(10),
+				retry.MaxDelay(10*time.Second),
+			)
+			require.NoError(t, err)
+			require.Equal(t, http.StatusBadRequest, status)
+		})
+
+	t.Run("server returns 200 when request uses HTTPS and valid password", func(t *testing.T) {
+		// Create a custom client with TLS configuration
+		client := &http.Client{
+			Transport: &http.Transport{
+				TLSClientConfig: &tls.Config{
+					InsecureSkipVerify: true,
+				},
+			},
+		}
+		status, err := retry.DoWithData(
+			func() (int, error) {
+				req := newRequestWithBasicAuth(t, "alice", "password", http.MethodGet, fmt.Sprintf("https://localhost:%d/metrics", port), nil)
+				resp, err := client.Do(req)
+				if err != nil {
+					return -1, err
+				}
+				return resp.StatusCode, nil
+			},
+			retry.Attempts(10),
+			retry.MaxDelay(10*time.Second),
+		)
+		require.NoError(t, err)
+		require.Equal(t, http.StatusOK, status)
+	})
+
+	t.Run("server returns 401 when request uses HTTPS and password is invalid", func(t *testing.T) {
+		// Create a custom client with TLS configuration
+		client := &http.Client{
+			Transport: &http.Transport{
+				TLSClientConfig: &tls.Config{
+					InsecureSkipVerify: true,
+				},
+			},
+		}
+		status, err := retry.DoWithData(
+			func() (int, error) {
+				req := newRequestWithBasicAuth(t, "alice", "bad password", http.MethodGet, fmt.Sprintf("https://localhost:%d/metrics", port), nil)
+				resp, err := client.Do(req)
+				if err != nil {
+					return -1, err
+				}
+				return resp.StatusCode, nil
+			},
+			retry.Attempts(10),
+			retry.MaxDelay(10*time.Second),
+		)
+		require.NoError(t, err)
+		require.Equal(t, http.StatusUnauthorized, status)
+	})
+	cancel()
+}
