+++ tests/integration/testdata/tlsCertificate.crt
@@ -1,0 +1,19 @@
+-----BEGIN CERTIFICATE-----
+MIIDHDCCAgSgAwIBAgIUGk7Hkcf3X3cTkf+AcO+LGnZ/gXcwDQYJKoZIhvcNAQEL
+BQAwEzERMA8GA1UEAwwIdmZlZG9yb3YwHhcNMjMxMjA2MTc0ODA5WhcNMzMxMjAz
+MTc0ODA5WjATMREwDwYDVQQDDAh2ZmVkb3JvdjCCASIwDQYJKoZIhvcNAQEBBQAD
+ggEPADCCAQoCggEBAM6eXKOZGnWWwHfLTWi86q6Fwtbpp+Kw3NGHkiZT9Fbevv+A
+7lEBHkisr/9Iya9BW4v01eabA6EUxFT93s8v0mb/TdSi1t2RczWStfQ9EVUoYHMw
+RBWughyEYBRmEgMs/9dYiSzYmAV4QjrKhyFf+9SyCd7C4Lc1S1WX2GRt+VEMvAdJ
+GBDhzAm1fJrsNq1FhcQd6skbsbRihVDYXcVZK8oy2Jzp1E/XVwxbrHgJpzoRmWtC
+lfv1piss+cP56epTb67fGg7Q/aDKVubc4FL2hrIt+iO6pMrzXQEZk0VJmdmXFLxk
+z1nb029jNqgxXaJcbQ+VEvs3r0yPyjm6PXOMzJMCAwEAAaNoMGYwHQYDVR0OBBYE
+FK8RnI5wqzc1gOmJrvVjC6k9L6QVMB8GA1UdIwQYMBaAFK8RnI5wqzc1gOmJrvVj
+C6k9L6QVMA8GA1UdEwEB/wQFMAMBAf8wEwYDVR0RBAwwCoIIdmZlZG9yb3YwDQYJ
+KoZIhvcNAQELBQADggEBAAA3QjlOZs3Aw8pgvw/DCDTr3ipO22d19LEzjUFlUKdj
+yDwmBE2O9Obebtjhv2R6ltkmaCsIa3li6DbpltfHLXq2bkzEckP7sbpq4QnJcAw/
+hjC65xli08HZO5VvtA9NYSZv6q8JdoY+r3pl8QtYTc4Wh/6D4lilv+OubQuGtjwE
+HXR4JPisNA70uLN5rUuMzKRTPmQCX0ljAuNmKKB7BHAxKxyGIJKVGMgcak+oEtR2
+wmEj+7Q4GkhYitTvJsBwm1XQyCuo+4nlZ2zHlELWklyIiuJ4YmUOBMOmgFtJVGRw
+ycp43YM3B4ZsOUNj+TXp7jG1isobfIjOlJrrlyHGbOk=
+-----END CERTIFICATE-----
