+++ tests/integration/testdata/tlsCertificate.key
@@ -1,0 +1,28 @@
+-----BEGIN PRIVATE KEY-----
+MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQDOnlyjmRp1lsB3
+y01ovOquhcLW6afisNzRh5ImU/RW3r7/gO5RAR5IrK//SMmvQVuL9NXmmwOhFMRU
+/d7PL9Jm/03UotbdkXM1krX0PRFVKGBzMEQVroIchGAUZhIDLP/XWIks2JgFeEI6
+yochX/vUsgnewuC3NUtVl9hkbflRDLwHSRgQ4cwJtXya7DatRYXEHerJG7G0YoVQ
+2F3FWSvKMtic6dRP11cMW6x4Cac6EZlrQpX79aYrLPnD+enqU2+u3xoO0P2gylbm
+3OBS9oayLfojuqTK810BGZNFSZnZlxS8ZM9Z29NvYzaoMV2iXG0PlRL7N69Mj8o5
+uj1zjMyTAgMBAAECggEAAp0UNOmR4RzoEFvc/c8t9MkyXr+T5RUxpeDeI2f6dkLZ
+DahW22V1OZwdmdAU5QehndKaBhG96QK3PQ/9CofjVHQ0StVIfIMDfxD9Bjrjhj09
+gxeANPoBIVU3Veq4nZurHNE84vlWavNLeyzmqrBpIhaic1PpvceJzWpfTV3qwnCt
+8s9BwCnYAQeZkG89H+o89veh8Gg8KnIQgeYuIBr8OSVPjpN2mx2/2HcmxMSLQzLd
+0WoGUqBe7INhsH8Ftr5Ft1esbVzVGuHYNpqzDON+ltkgHOXLsZZUFdZKWTvVcMl+
+DZJx60gLvS0gW1Pl+HUbfm9E+h7d87nydZHkyJqS6QKBgQD+4D1goP3sL7SGnZlD
+8y/y3EKulW9l5FluNcKaT8iAsb1CqFrH04bP9d6Dnf1FeugysUbq2+/DIuB4VgEH
+k2Ddd6uIOJqkFyQPFcueifjomsc5MnEsJUt4sK/ZuFFB9/MDberornS17s03GBlj
+bnQYcwwNwqay3kdDQSLH+4Jk7wKBgQDPh6N52ty4XQ9Kl3qchci35iC+0H4bbcOf
+m/QYmfhTWrAKuAbPKcDx3kmESIEZ39r1ne4piM2DpbIKRK5lp4zVCWC5+GjZeK5H
+rZTqNiPDCtQVPUgNrpq23lj98ZWgQlFbZmjD2+VpF1MhszgARN4nqtA6NxILYQkq
+K2+DhTV6nQKBgQCR1cbo0lVYXleCKwz2V2DZKnSxJmouOAF/0Rz/ZKVKFelACcqf
+ygW6Wmmuj6jKp6/SyaLVYXEIC5263SMLraVgM0gf0puRRnhodUUj+rhEh8N0k8sz
+wkcz0I1eG1vLTk9rZ+pgpum669/aJVzpitq83p9IBFeTwyGtdsiq+ts6WwKBgGKZ
+PfkqG83YIOi+mXl+MoP06T/fTNr5RoUxkkKR/KKVpqghENMiMlJ32xPnYrcJXT1u
+4HWJUqMmeTVZtkYD2expfTQ+KoqQRLU3alXwjQMgxOymZfJiL42MjFlyizVN7Ntx
+B3tuoUtR0qRjQssV7TqJdE8EVSVk3/bSFjQ1eNgdAoGAauCpps2W0yALIUWFRfDU
+RVOZWua079wbfvBvTNsQ0x9KDnfv7+3fZL7qoKyB+vLwuqgm1cj9IjyXDLppGVzJ
+Mk3zSbO9DwmOcM5IXnXwSd/zFFpUqH7FrGqUvfncMYgUrys1jmdY2yjr3iXRp/AT
+zP/AX7Fte5ABLuHZfxPkX6c=
+-----END PRIVATE KEY-----
